testing passes py2 + py3 with long paths
