+++ conans/paths.py
@@ -4,14 +4,42 @@ from os.path import join, normpath
 import platform
 from conans.errors import ConanException
 from conans.util.files import rmdir
+import sys
 
-if platform.system() == "Windows":
-    from conans.util.windows import path_shortener, rm_conandir, conan_expand_user
+
+def _check_long_paths_support():
+    if platform.system() != "Windows":
+        return True
+    if sys.version_info < (3, 6):
+        return False
+    from six.moves import winreg
+    try:
+        hKey = winreg.OpenKey(winreg.HKEY_LOCAL_MACHINE,
+                              r"SYSTEM\CurrentControlSet\Control\FileSystem")
+        result = winreg.QueryValueEx(hKey, "LongPathsEnabled")
+        key_value = result[0]
+        return key_value == 1
+    except EnvironmentError:
+        return False
+    finally:
+        winreg.CloseKey(hKey)
+    return False
+
+
+long_paths_support = _check_long_paths_support()
+
+if not long_paths_support:
+    from conans.util.windows import path_shortener, rm_conandir
 else:
     def path_shortener(x, _):
         return x
     conan_expand_user = os.path.expanduser
     rm_conandir = rmdir
+
+if platform.system() == "Windows":
+    from conans.util.windows import conan_expand_user
+else:
+    conan_expand_user = os.path.expanduser
 
 
 EXPORT_FOLDER = "export"
