+++ conans/util/windows.py
@@ -1,7 +1,6 @@
 import os
 from conans.util.files import load, mkdir, save, rmdir
 import tempfile
-import sys
 
 
 CONAN_LINK = ".conan_link"
@@ -30,78 +29,55 @@ def conan_expand_user(path):
     return result
 
 
-def _check_long_paths_support():
-    if sys.version_info < (3, 6):
-        return False
-    from six.moves import winreg
-    try:
-        hKey = winreg.OpenKey(winreg.HKEY_LOCAL_MACHINE,
-                              r"SYSTEM\CurrentControlSet\Control\FileSystem")
-        result = winreg.QueryValueEx(hKey, "LongPathsEnabled")
-        key_value = result[0]
-        return key_value == 1
-    except EnvironmentError:
-        return False
-    finally:
-        winreg.CloseKey(hKey)
-    return False
+def path_shortener(path, short_paths):
+    """ short_paths is 4-state:
+    False: Never shorten the path
+    True: Always shorten the path, create link if not existing
+    None: Use shorten path only if already exists, not create
+    """
+    if short_paths is False:
+        return path
+    link = os.path.join(path, CONAN_LINK)
+    if os.path.exists(link):
+        return load(link)
+    elif short_paths is None:
+        return path
+
+    short_home = os.getenv("CONAN_USER_HOME_SHORT")
+    if not short_home:
+        drive = os.path.splitdrive(path)[0]
+        short_home = drive + "/.conan"
+    mkdir(short_home)
+    redirect = tempfile.mkdtemp(dir=short_home, prefix="")
+    # This "1" is the way to have a non-existing directory, so commands like
+    # shutil.copytree() to it, works. It can be removed without compromising the
+    # temp folder generator and conan-links consistency
+    redirect = os.path.join(redirect, "1")
+    save(link, redirect)
+    return redirect
 
 
-if _check_long_paths_support():
-    def path_shortener(x, _):
-        return x
-    ignore_long_path_files = None
-    rm_conandir = rmdir
-else:
-    def path_shortener(path, short_paths):
-        """ short_paths is 4-state:
-        False: Never shorten the path
-        True: Always shorten the path, create link if not existing
-        None: Use shorten path only if already exists, not create
-        """
-        if short_paths is False or os.getenv("CONAN_USER_HOME_SHORT") == "None":
-            return path
-        link = os.path.join(path, CONAN_LINK)
-        if os.path.exists(link):
-            return load(link)
-        elif short_paths is None:
-            return path
-
-        short_home = os.getenv("CONAN_USER_HOME_SHORT")
-        if not short_home:
-            drive = os.path.splitdrive(path)[0]
-            short_home = drive + "/.conan"
-        mkdir(short_home)
-        redirect = tempfile.mkdtemp(dir=short_home, prefix="")
-        # This "1" is the way to have a non-existing directory, so commands like
-        # shutil.copytree() to it, works. It can be removed without compromising the
-        # temp folder generator and conan-links consistency
-        redirect = os.path.join(redirect, "1")
-        save(link, redirect)
-        return redirect
+def ignore_long_path_files(src_folder, build_folder, output):
+    def _filter(src, files):
+        filtered_files = []
+        for the_file in files:
+            source_path = os.path.join(src, the_file)
+            # Without storage path, just relative
+            rel_path = os.path.relpath(source_path, src_folder)
+            dest_path = os.path.normpath(os.path.join(build_folder, rel_path))
+            # it is NOT that "/" is counted as "\\" so it counts double
+            # seems a bug in python, overflows paths near the limit of 260,
+            if len(dest_path) >= 249:
+                filtered_files.append(the_file)
+                output.warn("Filename too long, file excluded: %s" % dest_path)
+        return filtered_files
+    return _filter
 
 
-    def ignore_long_path_files(src_folder, build_folder, output):
-        def _filter(src, files):
-            filtered_files = []
-            for the_file in files:
-                source_path = os.path.join(src, the_file)
-                # Without storage path, just relative
-                rel_path = os.path.relpath(source_path, src_folder)
-                dest_path = os.path.normpath(os.path.join(build_folder, rel_path))
-                # it is NOT that "/" is counted as "\\" so it counts double
-                # seems a bug in python, overflows paths near the limit of 260,
-                if len(dest_path) >= 249:
-                    filtered_files.append(the_file)
-                    output.warn("Filename too long, file excluded: %s" % dest_path)
-            return filtered_files
-        return _filter
-
-
-    def rm_conandir(path):
-        """removal of a directory that might contain a link to a short path"""
-        link = os.path.join(path, CONAN_LINK)
-        if os.path.exists(link):
-            short_path = load(link)
-            rmdir(os.path.dirname(short_path))
-        rmdir(path)
+def rm_conandir(path):
+    """removal of a directory that might contain a link to a short path"""
+    link = os.path.join(path, CONAN_LINK)
+    if os.path.exists(link):
+        short_path = load(link)
+        rmdir(os.path.dirname(short_path))
+    rmdir(path)
